6. Review: thay đổi hàm tạo invoice (DAO + View) , sua constrain customer_id thanh restrict + customerDAO

+++ module2/src/ra/phone_store_manager/business/impl/CustomerServiceImpl.java
@@ -59,6 +59,14 @@ public class CustomerServiceImpl implements ICustomerService {
             return false;
         }
 
+        // Kiểm tra khách hàng đã có hóa đơn chưa
+        if (customerDAO.checkCustomerHasInvoices(id)) {
+            System.out.println(Color.DO +
+                    "* Lỗi: Không thể xóa! Khách hàng này đã có lịch sử mua hàng."
+                    + Color.RESET);
+            return false;
+        }
+
         return customerDAO.deleteCustomer( id );
     }
